requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
gradio==4.16.0
requests==2.31.0
Pillow==10.1.0
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 